Converting Methods
Hi.

In a project we needed to convert C# interfaces into TypeScript, and one crucial part of this was that the methods of the interfaces were incldued as well..

I really liked the look of this project of yours and I took the liberty to add some basic functionality to include the methods as well.

- The methods are included in both classes and interfaces
- I added the new boolean option `methods` (default: false) that can be set to include methods.
- I added the new boolean option `returnPromise` (default false) to return the type as a promise from the generated method code
- Parameters with default values in the C# interface will become optional? in the TypeScript interface

This C# code:

```
namespace SomeInterfaces
{
  public interface ISomething
  {
       string name {get;set;}
  }

  public interface IConsumer
  {
       string ost {get;set;}
       void consumeIt(string value, ISomething apa);
       string ReturnSomething(string value = null);
   }

```

will become this ts-code (with the returnPromise option set to true):

```
declare module MyModule {
    // models\test.cs
    export interface ISomething {
        name: string;
    }

    // models\test.cs
    export interface IConsumer {
        ost: string;
        consumeIt (value: string, apa: ISomething): Promise<void>
        ReturnSomething (value?: string): Promise<string>
    }

}
```

I hope you think this is an ok addition to your project and that I have not messed up your code to badly.

/David