Added Directory.Build files to isolate cs projects
If you use the package in a project that has a Directory.Build.props or Directory.build.targets file in its root, the settings in those files control the compilation of the packages c# code. For instance, if nullable is enabled for the solution using the package, the csproj files will fail to build as they aren't nullable aware.